# cmd/task-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Масштабируемая микросервисная CRM-система на Go",
        "title": "CRM Distributed API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Торощина Екатерина",
            "email": "katya@crm.ru"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/projects/{uuid}/tasks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает задачи проекта с фильтрацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Список задач проекта",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID проекта",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Фильтр по статусу",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Поиск по названию",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tasks": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт новую задачу в проекте с атомарным счётчиком (FOR UPDATE)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Создать задачу",
                "parameters": [
                    {
                        "description": "Параметры задачи",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/crm-distributed_shared_domain.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{uuid}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает задачу по UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получить задачу",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID задачи",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/crm-distributed_shared_domain.Task"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Мягкое удаление задачи (soft delete)",
                "tags": [
                    "tasks"
                ],
                "summary": "Удалить задачу",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID задачи",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{uuid}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Переводит задачу в новый статус по статусному графу",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Изменить статус задачи",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID задачи",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый статус",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "crm-distributed_shared_domain.Activity": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "$ref": "#/definitions/crm-distributed_shared_domain.User"
                },
                "description": {
                    "type": "string"
                },
                "entityType": {
                    "type": "string"
                },
                "entityUUID": {
                    "type": "string"
                },
                "meta": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "type": {
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "crm-distributed_shared_domain.ProfilePhotoDTO": {
            "type": "object",
            "properties": {
                "large": {
                    "type": "string"
                },
                "medium": {
                    "type": "string"
                },
                "small": {
                    "type": "string"
                }
            }
        },
        "crm-distributed_shared_domain.ProfilePreferences": {
            "type": "object",
            "properties": {
                "timezone": {
                    "type": "string"
                }
            }
        },
...
        "crm-distributed_shared_domain.Task": {
            "type": "object",
            "required": [
                "companyUUID",
                "federationUUID",
                "projectUUID"
            ],
            "properties": {
                "activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/crm-distributed_shared_domain.Activity"
                    }
                },
                "activitiesTotal": {
                    "type": "integer",
                    "format": "int64"
                },
                "activityAt": {
                    "type": "string"
                },
                "childrenTotal": {
                    "type": "integer"
                },
                "childrenUUIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "coWorkersBy": {
                    "description": "соисполнитель и наблюдатель",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "commentsTotal": {
                    "type": "integer"
                },
                "companyUUID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "description": "роли на задаче",
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "dirty": {
                    "description": "для постройки изменений activity",
                    "type": "object",
                    "additionalProperties": {}
                },
                "federationUUID": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "finishTo": {
                    "description": "дедлайн",
                    "type": "string"
                },
                "finishedAt": {
                    "type": "string"
                },
                "finishedBy": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "implementBy": {
                    "type": "string"
                },
                "isEpic": {
                    "description": "задача является эпиком(контейнером для подзадач)",
                    "type": "boolean"
                },
                "managedBy": {
                    "type": "string"
                },
                "meta": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "path": {
                    "description": "иерархмческий путь для поддержки подзадач",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "people": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "priority": {
                    "type": "integer"
                },
                "projectUUID": {
                    "type": "string"
                },
                "rawFields": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "responsibleBy": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "watchBy": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "crm-distributed_shared_domain.User": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "hasPhoto": {
                    "type": "boolean"
                },
                "isValid": {
                    "type": "boolean"
                },
                "lname": {
                    "type": "string",
                    "maxLength": 30
                },
                "meta": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 30
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "integer",
                    "format": "int64"
                },
                "photo": {
                    "$ref": "#/definitions/crm-distributed_shared_domain.ProfilePhotoDTO"
                },
                "pname": {
                    "type": "string",
                    "maxLength": 30
                },
                "preferences": {
                    "$ref": "#/definitions/crm-distributed_shared_domain.ProfilePreferences"
                },
                "provider": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "validAt": {
                    "type": "string"
                },
                "validationSendAt": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}